move SearchIndexRefresher to the jobs package
